headless-requirements.txt
tensorflow_addons==0.18.0
matplotlib
scipy
scikit-image
ipython
tqdm
tensorflowjs==3.21.0
tensorflow-io==0.27.0